requirements.txt
Flask==2.2.2
requests==2.28.2
gspread==5.7.2
google-auth==2.16.2
beautifulsoup4==4.11.1
